delete book admin password validaiton with express-validator, fix add book bug

+++ controllers/indexController.js
@@ -1,4 +1,5 @@
 const db = require("../db/queries");
+const { body, validationResult } = require("express-validator");
 
 async function indexGet(req, res) {
   const { genreName } = req.query;
@@ -9,9 +10,7 @@ async function indexGet(req, res) {
   if (genreName && genreName !== "") {
     books = await db.getBooksByGenre(genreName);
     selectedGenre = genreName;
-  } 
-  
-  else {
+  } else {
     books = await db.getAllBooks();
   }
 
@@ -28,21 +27,45 @@ async function bookDetailsGet(req, res) {
 }
 
 async function bookDeleteGet(req, res) {
-    const bookId = req.params.bookId;
-    const book = await db.getBookById(bookId);
-    const bookTitle = book.title
-    res.render("deleteBook", { bookId, bookTitle })
+  const bookId = req.params.bookId;
+  const book = await db.getBookById(bookId);
+  const bookTitle = book.title;
+  res.render("deleteBook", { bookId, bookTitle });
 }
 
-async function bookDeletePost(req, res) {
-  const bookId = req.params.bookId;
-  await db.deleteBookById(bookId);
-  res.redirect("/");
-}
+const bookDeletePost = [
+  body("admin_password")
+    .trim()
+    .notEmpty().withMessage("Password is required.")
+    .custom((value) => {
+      if (value !== process.env.ADMIN_PASSWORD) {
+        throw new Error("Wrong password. Please try again.");
+      }
+      return true;
+    }),
+
+  async (req, res) => {
+    const bookId = req.params.bookId;
+    const book = await db.getBookById(bookId);
+    const bookTitle = book.title;
+    const errors = validationResult(req);
+
+    if (!errors.isEmpty()) {
+      return res.render("deleteBook", {
+        bookId,
+        bookTitle,
+        errors: errors.array(),
+      });
+    }
+
+    await db.deleteBookById(bookId);
+    res.redirect("/");
+  },
+];
 
 module.exports = {
   indexGet,
   bookDetailsGet,
   bookDeletePost,
-  bookDeleteGet
+  bookDeleteGet,
 };
